requirements.txt
Flask>=2.0
reportlab>=3.5
python-dotenv>=0.19